+++ client/components/index.js
@@ -10,4 +10,3 @@ export {default as AllProducts} from './allProducts'
 export {default as SingleProduct} from './singleProduct'
 export {default as ShoppingCart} from './shoppingCart'
 export {default as Profile} from './profile'
-export {default as OrderHistory} from './orderHistory'
